Don't fail if no value is returned by _get
In case something isn't right with the `_get` method, we should avoid
throwing an error when trying to call `.toString()` on `undefined`.

You could argue that the absence of a proper `value` is in it self an
error, which is true, but when tape throws an error like this it masks
any assertion errors that might have have occurred before this error was
thrown. This makes it extremely difficult to debug why the error
occurred in the first place as there might have been a meaningfull
assertion error just a couple of lines before that would have told you
what was wrong.