src: Allow buffer / nextTick to be overwritten
There were two dependencies added. Buffer and immediate.

This is for browser support. Instead allow references to Buffer
and nextTick to be overwritten by the sub class.

This allows the browser module level-js to import these modules
and overwrite the methods using them.

Also in this PR inline xtend and remove it as a dependency.